extract mount handling bits out of AptConfigurer
