+++ subiquity/server/apt.py
@@ -14,87 +14,27 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import contextlib
-import functools
 import logging
 import os
 import shutil
 import subprocess
-import tempfile
-from typing import List, Optional, Union
-
-import attr
+from typing import Optional
 
 from curtin.config import merge_config
 
 from subiquitycore.file_util import write_file, generate_config_yaml
 from subiquitycore.lsb_release import lsb_release
-from subiquitycore.utils import arun_command
 
 from subiquity.server.curtin import run_curtin_command
+from subiquity.server.mounter import (
+    DryRunMounter,
+    Mounter,
+    Mountpoint,
+    OverlayCleanupError,
+    OverlayMountpoint,
+    )
 
 log = logging.getLogger('subiquity.server.apt')
-
-
-class OverlayCleanupError(Exception):
-    """ Exception to raise when an overlay could not be cleaned up. """
-
-
-class _MountBase:
-
-    def p(self, *args: str) -> str:
-        for a in args:
-            if a.startswith('/'):
-                raise Exception('no absolute paths here please')
-        return os.path.join(self.mountpoint, *args)
-
-    def write(self, path, content):
-        with open(self.p(path), 'w') as fp:
-            fp.write(content)
-
-
-@attr.s(auto_attribs=True, kw_only=True)
-class Mountpoint(_MountBase):
-    mountpoint: str
-
-
-@attr.s(auto_attribs=True, kw_only=True)
-class OverlayMountpoint(_MountBase):
-    # The first element in lowers will be the bottom layer and the last element
-    # will be the top layer.
-    lowers: List["Lower"]
-    upperdir: Optional[str]
-    mountpoint: str
-
-
-Lower = Union[Mountpoint, str, OverlayMountpoint]
-
-
-@functools.singledispatch
-def lowerdir_for(x):
-    """Return value suitable for passing to the lowerdir= overlayfs option."""
-    raise NotImplementedError(x)
-
-
-@lowerdir_for.register(str)
-def _lowerdir_for_str(path):
-    return path
-
-
-@lowerdir_for.register(Mountpoint)
-def _lowerdir_for_mnt(mnt):
-    return mnt.mountpoint
-
-
-@lowerdir_for.register(OverlayMountpoint)
-def _lowerdir_for_ovmnt(ovmnt):
-    # One cannot indefinitely stack overlayfses so construct an
-    # explicit list of the layers of the overlayfs.
-    return lowerdir_for([ovmnt.lowers, ovmnt.upperdir])
-
-
-@lowerdir_for.register(list)
-def _lowerdir_for_lst(lst):
-    return ':'.join(reversed([lowerdir_for(item) for item in lst]))
 
 
 class AptConfigurer:
@@ -133,57 +73,13 @@ class AptConfigurer:
     #    system, or if it is not, just copy /var/lib/apt/lists from the
     #    'configured_tree' overlay.
 
-    def __init__(self, app, source: str):
+    def __init__(self, app, mounter: Mounter, source: str):
         self.app = app
+        self.mounter = mounter
         self.source: str = source
         self.configured_tree: Optional[OverlayMountpoint] = None
         self.install_tree: Optional[OverlayMountpoint] = None
         self.install_mount = None
-        self._mounts: List[Mountpoint] = []
-        self._tdirs: List[str] = []
-
-    def tdir(self):
-        d = tempfile.mkdtemp()
-        self._tdirs.append(d)
-        return d
-
-    async def mount(self, device, mountpoint, options=None, type=None):
-        opts = []
-        if options is not None:
-            opts.extend(['-o', options])
-        if type is not None:
-            opts.extend(['-t', type])
-        await self.app.command_runner.run(
-            ['mount'] + opts + [device, mountpoint], private_mounts=False)
-        m = Mountpoint(mountpoint=mountpoint)
-        self._mounts.append(m)
-        return m
-
-    async def unmount(self, mountpoint: Mountpoint, remove=True):
-        if remove:
-            self._mounts.remove(mountpoint)
-        await self.app.command_runner.run(
-                ['umount', mountpoint.mountpoint],
-                private_mounts=False)
-
-    async def setup_overlay(self, lowers: List[Lower]) -> OverlayMountpoint:
-        tdir = self.tdir()
-        target = f'{tdir}/mount'
-        lowerdir = lowerdir_for(lowers)
-        upperdir = f'{tdir}/upper'
-        workdir = f'{tdir}/work'
-        for d in target, workdir, upperdir:
-            os.mkdir(d)
-
-        options = f'lowerdir={lowerdir},upperdir={upperdir},workdir={workdir}'
-
-        mount = await self.mount(
-            'overlay', target, options=options, type='overlay')
-
-        return OverlayMountpoint(
-            lowers=lowers,
-            mountpoint=mount.p(),
-            upperdir=upperdir)
 
     def apt_config(self):
         cfg = {}
@@ -192,7 +88,7 @@ class AptConfigurer:
         return {'apt': cfg}
 
     async def apply_apt_config(self, context):
-        self.configured_tree = await self.setup_overlay([self.source])
+        self.configured_tree = await self.mounter.setup_overlay([self.source])
 
         config_location = os.path.join(
             self.app.root, 'var/log/installer/subiquity-curtin-apt.conf')
@@ -206,10 +102,11 @@ class AptConfigurer:
     async def configure_for_install(self, context):
         assert self.configured_tree is not None
 
-        self.install_tree = await self.setup_overlay([self.configured_tree])
+        self.install_tree = await self.mounter.setup_overlay(
+            [self.configured_tree])
 
         os.mkdir(self.install_tree.p('cdrom'))
-        await self.mount(
+        await self.mounter.mount(
             '/cdrom', self.install_tree.p('cdrom'), options='bind')
 
         if self.app.base_model.network.has_network:
@@ -236,7 +133,7 @@ class AptConfigurer:
 
     @contextlib.asynccontextmanager
     async def overlay(self):
-        overlay = await self.setup_overlay([
+        overlay = await self.mounter.setup_overlay([
                 self.install_tree.upperdir,
                 self.configured_tree.upperdir,
                 self.source
@@ -251,18 +148,13 @@ class AptConfigurer:
             # compares equal to the one we discarded earlier.
             # But really, there should be better ways to handle this.
             try:
-                await self.unmount(Mountpoint(mountpoint=overlay.mountpoint))
+                await self.mounter.unmount(
+                    Mountpoint(mountpoint=overlay.mountpoint))
             except subprocess.CalledProcessError as exc:
                 raise OverlayCleanupError from exc
 
     async def cleanup(self):
-        for m in reversed(self._mounts):
-            await self.unmount(m, remove=False)
-        for d in self._tdirs:
-            try:
-                shutil.rmtree(d)
-            except OSError as ose:
-                log.warning(f'failed to rmtree {d}: {ose}')
+        await self.mounter.cleanup()
 
     async def deconfigure(self, context, target: str) -> None:
         target_mnt = Mountpoint(mountpoint=target)
@@ -293,43 +185,15 @@ class AptConfigurer:
         # Call this after the rootfs has been extracted to the real target
         # system but before any configuration is applied to it.
         target_mnt = Mountpoint(mountpoint=target)
-        await self.mount(
+        await self.mounter.mount(
             '/cdrom', target_mnt.p('cdrom'), options='bind')
 
 
 class DryRunAptConfigurer(AptConfigurer):
 
-    async def unmount(self, mountpoint: Mountpoint, remove=True):
-        if remove:
-            self._mounts.remove(mountpoint)
-
-    async def setup_overlay(self, lowers: List[Lower]) -> OverlayMountpoint:
-        # XXX This implementation expects that:
-        # - on first invocation, the lowers list contains a single string
-        # element.
-        # - on second invocation, the lowers list contains the
-        # OverlayMountPoint returned by the first invocation.
-        lowerdir = lowers[0]
-        if isinstance(lowerdir, OverlayMountpoint):
-            source = lowerdir.lowers[0]
-        else:
-            source = lowerdir
-        target = self.tdir()
-        os.mkdir(f'{target}/etc')
-        await arun_command([
-            'cp', '-aT', f'{source}/etc/apt', f'{target}/etc/apt',
-            ], check=True)
-        if os.path.isdir(f'{target}/etc/apt/sources.list.d'):
-            shutil.rmtree(f'{target}/etc/apt/sources.list.d')
-        os.mkdir(f'{target}/etc/apt/sources.list.d')
-        return OverlayMountpoint(
-            lowers=[source],
-            mountpoint=target,
-            upperdir=None)
-
     @contextlib.asynccontextmanager
     async def overlay(self):
-        yield await self.setup_overlay(self.install_tree.mountpoint)
+        yield await self.mounter.setup_overlay(self.install_tree.mountpoint)
 
     async def deconfigure(self, context, target):
         await self.cleanup()
@@ -337,6 +201,6 @@ class DryRunAptConfigurer(AptConfigurer):
 
 def get_apt_configurer(app, source: str):
     if app.opts.dry_run:
-        return DryRunAptConfigurer(app, source)
+        return DryRunAptConfigurer(app, DryRunMounter(app), source)
     else:
-        return AptConfigurer(app, source)
+        return AptConfigurer(app, Mounter(app), source)
