+++ subiquity/server/tests/test_apt.py
@@ -19,9 +19,7 @@ from subiquitycore.tests import SubiTestCase
 from subiquitycore.tests.mocks import make_app
 from subiquity.server.apt import (
     AptConfigurer,
-    Mountpoint,
     OverlayMountpoint,
-    lowerdir_for,
 )
 from subiquity.models.mirror import MirrorModel, DEFAULT
 from subiquity.models.proxy import ProxyModel
@@ -36,7 +34,7 @@ class TestAptConfigurer(SubiTestCase):
         self.model.mirror = MirrorModel()
         self.model.proxy = ProxyModel()
         self.app = make_app(self.model)
-        self.configurer = AptConfigurer(self.app, '')
+        self.configurer = AptConfigurer(self.app, AsyncMock(), '')
 
     def test_apt_config_noproxy(self):
         expected = self.base_config
@@ -49,13 +47,6 @@ class TestAptConfigurer(SubiTestCase):
         expected['apt']['http_proxy'] = proxy
         expected['apt']['https_proxy'] = proxy
         self.assertEqual(expected, self.configurer.apt_config())
-
-    async def test_mount_unmount(self):
-        # Make sure we can unmount something that we mounted before.
-        with patch.object(self.app, "command_runner",
-                          create=True, new_callable=AsyncMock):
-            m = await self.configurer.mount("/dev/cdrom", "/target")
-            await self.configurer.unmount(m)
 
     async def test_overlay(self):
         self.configurer.install_tree = OverlayMountpoint(
@@ -74,59 +65,3 @@ class TestAptConfigurer(SubiTestCase):
                           create=True, new_callable=AsyncMock):
             async with self.configurer.overlay():
                 pass
-
-
-class TestLowerDirFor(SubiTestCase):
-    def test_lowerdir_for_str(self):
-        self.assertEqual(
-                lowerdir_for("/tmp/lower1"),
-                "/tmp/lower1")
-
-    def test_lowerdir_for_str_list(self):
-        self.assertEqual(
-                lowerdir_for(["/tmp/lower1", "/tmp/lower2"]),
-                "/tmp/lower2:/tmp/lower1")
-
-    def test_lowerdir_for_mountpoint(self):
-        self.assertEqual(
-                lowerdir_for(Mountpoint(mountpoint="/mnt")),
-                "/mnt")
-
-    def test_lowerdir_for_simple_overlay(self):
-        overlay = OverlayMountpoint(
-                lowers=["/tmp/lower1"],
-                upperdir="/tmp/upper1",
-                mountpoint="/mnt",
-        )
-        self.assertEqual(lowerdir_for(overlay), "/tmp/upper1:/tmp/lower1")
-
-    def test_lowerdir_for_overlay(self):
-        overlay = OverlayMountpoint(
-                lowers=["/tmp/lower1", "/tmp/lower2"],
-                upperdir="/tmp/upper1",
-                mountpoint="/mnt",
-        )
-        self.assertEqual(
-                lowerdir_for(overlay),
-                "/tmp/upper1:/tmp/lower2:/tmp/lower1")
-
-    def test_lowerdir_for_list(self):
-        overlay = OverlayMountpoint(
-                lowers=["/tmp/overlaylower1", "/tmp/overlaylower2"],
-                upperdir="/tmp/overlayupper1",
-                mountpoint="/mnt/overlay",
-        )
-        mountpoint = Mountpoint(mountpoint="/mnt/mountpoint")
-        lowers = ["/tmp/lower1", "/tmp/lower2"]
-        self.assertEqual(
-                lowerdir_for([overlay, mountpoint, lowers]),
-                "/tmp/lower2:/tmp/lower1" +
-                ":/mnt/mountpoint" +
-                ":/tmp/overlayupper1:/tmp/overlaylower2:/tmp/overlaylower1")
-
-    def test_lowerdir_for_other(self):
-        with self.assertRaises(NotImplementedError):
-            lowerdir_for(None)
-
-        with self.assertRaises(NotImplementedError):
-            lowerdir_for(10)
